+++ tests/vector_stores/test_postgres.py
@@ -1,10 +1,11 @@
 import asyncio
-from typing import Any, Dict, Generator, List, Union
+from typing import List, Any, Dict, Union, Generator, cast
 
 import pytest
 
 from llama_index.schema import NodeRelationship, RelatedNodeInfo, TextNode
 from llama_index.vector_stores import PGVectorStore
+from llama_index.vector_stores.loading import load_vector_store
 from llama_index.vector_stores.types import (
     ExactMatchFilter,
     MetadataFilters,
@@ -17,7 +18,7 @@ from llama_index.vector_stores.types import (
 
 
 PARAMS: Dict[str, Union[str, int]] = dict(
-    host="localhost", user="postgres", password="password", port=5432
+    host="localhost", user="postgres", password="mark90", port=5432
 )
 TEST_DB = "test_vector_db"
 TEST_TABLE_NAME = "lorem_ipsum"
@@ -242,7 +243,21 @@ async def test_add_to_db_query_and_delete(
     assert res.nodes
     assert len(res.nodes) == 1
     assert res.nodes[0].node_id == "bbb"
-    pg.delete("bbb")
+
+
+@pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
+@pytest.mark.asyncio
+@pytest.mark.parametrize("use_async", [(True,), (False,)])
+async def test_save_load(
+    pg: PGVectorStore, node_embeddings: List[NodeWithEmbedding], use_async: bool
+) -> None:
+    if use_async:
+        await pg.async_add(node_embeddings)
+    else:
+        pg.add(node_embeddings)
+    assert isinstance(pg, PGVectorStore)
+
+    q = VectorStoreQuery(query_embedding=_get_sample_vector(0.1), similarity_top_k=1)
 
     if use_async:
         res = await pg.aquery(q)
@@ -250,7 +265,25 @@ async def test_add_to_db_query_and_delete(
         res = pg.query(q)
     assert res.nodes
     assert len(res.nodes) == 1
-    assert res.nodes[0].node_id == "aaa"
+    assert res.nodes[0].node_id == "bbb"
+
+    pg_dict = pg.to_dict()
+    await pg.close()
+
+    loaded_pg = cast(PGVectorStore, load_vector_store(pg_dict))
+    loaded_pg_dict = loaded_pg.to_dict()
+    for key, val in pg.to_dict().items():
+        assert loaded_pg_dict[key] == val
+
+    if use_async:
+        res = await loaded_pg.aquery(q)
+    else:
+        res = loaded_pg.query(q)
+    assert res.nodes
+    assert len(res.nodes) == 1
+    assert res.nodes[0].node_id == "bbb"
+
+    await loaded_pg.close()
 
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")