+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## Unreleased
+
+### New Features
+- Added support for Azure Cognitive Search vector store (#7469)
+- Support delete in supabase (#6951)
+
 ## [0.8.20] - 2023-09-04
 
 ### New Features